+++ templates/init/functions/javascript/index.js
